# src/swagger/product.yaml
openapi: 3.0.0
info:
  title: Multi-Vendor eCommerce API
  version: 1.0.0
  description: API for managing products in a multi-vendor eCommerce platform.
servers:
  - url: "http://localhost:5000"
    description: Local server
paths:
...
  /api/products/{id}:
    get:
      summary: Get a specific product
      description: Retrieves the details of a single product by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The product's ID
          schema:
            type: string
      responses:
        200:
          description: Product details
        404:
          description: Product not found
    put:
      summary: Update a product
      description: Allows a seller to update an existing product by ID with new data.
      parameters:
        - name: id
          in: path
          required: true
          description: The product's ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Updated Product Name"
                price:
                  type: number
                  example: 30.00
                description:
                  type: string
                  example: "Updated product description"
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: "Updated images for the product"
      responses:
        200:
          description: Product updated successfully
        400:
          description: Invalid data provided
    delete:
      summary: Delete a product
      description: Allows a seller to delete a product by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The product's ID
          schema:
            type: string
      responses:
        204:
          description: Product deleted successfully
        404:
          description: Product not found